requirements.txt
bitbucket-pipes-toolkit~=3.2.1
GitPython~=3.1.14
requests~=2.25.1
packaging~=21.0